llm/udemy-aiagents-workflows-practical-guide/03-ai-agents/pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.91.0",
    "python-dotenv>=1.1.1",
]
